Move ratio_from_viewbox to SVGElementData
