+++ components/shared/layout/lib.rs
@@ -45,7 +45,7 @@ use servo_arc::Arc as ServoArc;
 use servo_url::{ImmutableOrigin, ServoUrl};
 use style::Atom;
 use style::animation::DocumentAnimationSet;
-use style::attr::AttrValue;
+use style::attr::{AttrValue, parse_integer, parse_unsigned_integer};
 use style::context::QuirksMode;
 use style::data::ElementData;
 use style::dom::OpaqueNode;
@@ -54,6 +54,7 @@ use style::media_queries::Device;
 use style::properties::style_structs::Font;
 use style::properties::{ComputedValues, PropertyId};
 use style::selector_parser::{PseudoElement, RestyleDamage, Snapshot};
+use style::str::char_is_whitespace;
 use style::stylesheets::{Stylesheet, UrlExtraData};
 use style::values::computed::Overflow;
 use style_traits::CSSPixel;
@@ -139,6 +140,27 @@ pub struct SVGElementData<'dom> {
     pub width: Option<&'dom AttrValue>,
     pub height: Option<&'dom AttrValue>,
     pub view_box: Option<&'dom AttrValue>,
+}
+
+impl SVGElementData<'_> {
+    pub fn ratio_from_view_box(view_box: &AttrValue) -> Option<f32> {
+        let mut iter = view_box.chars();
+        let _min_x = parse_integer(&mut iter).ok()?;
+        let _min_y = parse_integer(&mut iter).ok()?;
+
+        let width = parse_unsigned_integer(&mut iter).ok()?;
+        if width == 0 {
+            return None;
+        }
+
+        let height = parse_unsigned_integer(&mut iter).ok()?;
+        if height == 0 {
+            return None;
+        }
+
+        let mut iter = iter.skip_while(|c| char_is_whitespace(*c));
+        iter.next().is_none().then(|| width as f32 / height as f32)
+    }
 }
 
 /// The address of a node known to be valid. These are sent from script to layout.
